highlight target locations
When following a link (especially if the target is small, e.g. a backlink to
a footnote, or if the target doesn't appear right at the top of the page)
it can be unclear what the actual target is.
We highlight with a yellow flash (for footnotes) or electric blue
(for TOC and other internal links).